requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Marshmallow==0.15.0
marshmallow==3.19.0
psycopg2-binary==2.9.7
